Move the implementation of ComputeIndexOwner into mpi.cc
